+++ src/components/AboutSection.tsx
@@ -1,0 +1,19 @@
+export function AboutSection() {
+  return (
+    <>
+      <h2 className="text-xl font-bold mb-2">About Me</h2>
+      <div className="flex gap-6 items-center">
+        <p className="text-sm text-gray-500 dark:text-gray-300">
+          Engineering professional with 15+ years of experience delivering
+          scalable, secure back-end systems using PHP, Laravel, and modern
+          DevOps practices. Proven success leading Laravel integrations, CI/CD
+          pipelines, and infrastructure improvements in high-stakes environments
+          (banking, e-commerce, education). Strong communicator and mentor with
+          a deep understanding of agile workflows, technical debt reduction, and
+          legacy system refactoring. Looking to contribute in a Lead Engineer
+          capacity within a collaborative, fast-moving hybrid team.
+        </p>
+      </div>
+    </>
+  );
+}
